Initial work to decouple Editor from SWT/Eclipse
